Simplify the implementation
